move parser to separate folder
new folder name is 'dashboard_parser'

importParser renamed to parser.py
importParserHelper renamed to parser_helper.py
